backend/requirements.txt
# AI翻译助手 - 后端依赖

# Web框架
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# HTTP客户端
httpx>=0.27.0

# 数据验证
pydantic>=2.10.0

# 环境变量管理
python-dotenv>=1.0.0

